# feat: optimise generic decodes