--- a/codec_native.go
+++ b/codec_native.go
@@ -12,6 +12,7 @@ import (
 
 //nolint:maintidx // Splitting this would not make it simpler.
 func createDecoderOfNative(schema *PrimitiveSchema, typ reflect2.Type) ValDecoder {
+	isConv := schema.actual != ""
 	switch typ.Kind() {
 	case reflect.Bool:
 		if schema.Type() != Boolean {
@@ -59,7 +60,10 @@ func createDecoderOfNative(schema *PrimitiveSchema, typ reflect2.Type) ValDecode
 		if schema.Type() != Long {
 			break
 		}
-		return &longCodec[uint32]{convert: createLongConverter(schema.actual)}
+		if isConv {
+			return &longConvCodec[uint32]{convert: createLongConverter(schema.actual)}
+		}
+		return &longCodec[uint32]{}
 
 	case reflect.Int64:
 		st := schema.Type()
@@ -74,9 +78,10 @@ func createDecoderOfNative(schema *PrimitiveSchema, typ reflect2.Type) ValDecode
 			}
 
 		case st == Long:
-			return &longCodec[int64]{
-				convert: createLongConverter(schema.actual),
+			if isConv {
+				return &longConvCodec[int64]{convert: createLongConverter(schema.actual)}
 			}
+			return &longCodec[int64]{}
 
 		default:
 			break
@@ -86,34 +91,31 @@ func createDecoderOfNative(schema *PrimitiveSchema, typ reflect2.Type) ValDecode
 		if schema.Type() != Float {
 			break
 		}
-		return &float32Codec{
-			convert: createFloatConverter(schema.actual),
+		if isConv {
+			return &float32ConvCodec{convert: createFloatConverter(schema.actual)}
 		}
+		return &float32Codec{}
 
 	case reflect.Float64:
 		if schema.Type() != Double {
 			break
 		}
-		return &float64Codec{
-			convert: createDoubleConverter(schema.actual),
+		if isConv {
+			return &float64ConvCodec{convert: createDoubleConverter(schema.actual)}
 		}
+		return &float64Codec{}
 
 	case reflect.String:
 		if schema.Type() != String {
 			break
 		}
-		return &stringCodec{
-			convert: createStringConverter(schema.actual),
-		}
+		return &stringCodec{}
 
 	case reflect.Slice:
 		if typ.(reflect2.SliceType).Elem().Kind() != reflect.Uint8 || schema.Type() != Bytes {
 			break
 		}
-		return &bytesCodec{
-			sliceType: typ.(*reflect2.UnsafeSliceType),
-			convert:   createBytesConverter(schema.actual),
-		}
+		return &bytesCodec{sliceType: typ.(*reflect2.UnsafeSliceType)}
 
 	case reflect.Struct:
 		st := schema.Type()
@@ -143,10 +145,7 @@ func createDecoderOfNative(schema *PrimitiveSchema, typ reflect2.Type) ValDecode
 			}
 		case typ.Type1().ConvertibleTo(ratType) && st == Bytes && lt == Decimal:
 			dec := ls.(*DecimalLogicalSchema)
-			return &bytesDecimalCodec{
-				prec: dec.Precision(), scale: dec.Scale(),
-				convert: createBytesConverter(schema.actual),
-			}
+			return &bytesDecimalCodec{prec: dec.Precision(), scale: dec.Scale()}
 
 		default:
 			break
@@ -164,10 +163,7 @@ func createDecoderOfNative(schema *PrimitiveSchema, typ reflect2.Type) ValDecode
 		}
 		dec := ls.(*DecimalLogicalSchema)
 
-		return &bytesDecimalPtrCodec{
-			prec: dec.Precision(), scale: dec.Scale(),
-			convert: createBytesConverter(schema.actual),
-		}
+		return &bytesDecimalPtrCodec{prec: dec.Precision(), scale: dec.Scale()}
 	}
 
 	return &errorDecoder{err: fmt.Errorf("avro: %s is unsupported for Avro %s", typ.String(), schema.Type())}
@@ -362,41 +358,40 @@ type largeInt interface {
 	~int32 | ~uint32 | int64
 }
 
-type longCodec[T largeInt] struct {
-	convert func(*Reader) int64
-}
+type longCodec[T largeInt] struct{}
 
 func (c *longCodec[T]) Decode(ptr unsafe.Pointer, r *Reader) {
-	var v T
-	if c.convert != nil {
-		v = T(c.convert(r))
-	} else {
-		v = T(r.ReadLong())
-	}
-	*((*T)(ptr)) = v
+	*((*T)(ptr)) = T(r.ReadLong())
 }
 
 func (*longCodec[T]) Encode(ptr unsafe.Pointer, w *Writer) {
 	w.WriteLong(int64(*((*T)(ptr))))
 }
 
-type float32Codec struct {
-	convert func(*Reader) float32
+type longConvCodec[T largeInt] struct {
+	convert func(*Reader) int64
 }
 
-func (c *float32Codec) Decode(ptr unsafe.Pointer, r *Reader) {
-	var v float32
-	if c.convert != nil {
-		v = c.convert(r)
-	} else {
-		v = r.ReadFloat()
-	}
+func (c *longConvCodec[T]) Decode(ptr unsafe.Pointer, r *Reader) {
+	*((*T)(ptr)) = T(c.convert(r))
+}
 
-	*((*float32)(ptr)) = v
+type float32Codec struct{}
+
+func (c *float32Codec) Decode(ptr unsafe.Pointer, r *Reader) {
+	*((*float32)(ptr)) = r.ReadFloat()
 }
 
 func (*float32Codec) Encode(ptr unsafe.Pointer, w *Writer) {
 	w.WriteFloat(*((*float32)(ptr)))
+}
+
+type float32ConvCodec struct {
+	convert func(*Reader) float32
+}
+
+func (c *float32ConvCodec) Decode(ptr unsafe.Pointer, r *Reader) {
+	*((*float32)(ptr)) = c.convert(r)
 }
 
 type float32DoubleCodec struct{}
@@ -405,22 +400,22 @@ func (*float32DoubleCodec) Encode(ptr unsafe.Pointer, w *Writer) {
 	w.WriteDouble(float64(*((*float32)(ptr))))
 }
 
-type float64Codec struct {
-	convert func(*Reader) float64
-}
+type float64Codec struct{}
 
 func (c *float64Codec) Decode(ptr unsafe.Pointer, r *Reader) {
-	var v float64
-	if c.convert != nil {
-		v = c.convert(r)
-	} else {
-		v = r.ReadDouble()
-	}
-	*((*float64)(ptr)) = v
+	*((*float64)(ptr)) = r.ReadDouble()
 }
 
 func (*float64Codec) Encode(ptr unsafe.Pointer, w *Writer) {
 	w.WriteDouble(*((*float64)(ptr)))
+}
+
+type float64ConvCodec struct {
+	convert func(*Reader) float64
+}
+
+func (c *float64ConvCodec) Decode(ptr unsafe.Pointer, r *Reader) {
+	*((*float64)(ptr)) = c.convert(r)
 }
 
 type stringCodec struct {
@@ -428,13 +423,7 @@ type stringCodec struct {
 }
 
 func (c *stringCodec) Decode(ptr unsafe.Pointer, r *Reader) {
-	var v string
-	if c.convert != nil {
-		v = c.convert(r)
-	} else {
-		v = r.ReadString()
-	}
-	*((*string)(ptr)) = v
+	*((*string)(ptr)) = r.ReadString()
 }
 
 func (*stringCodec) Encode(ptr unsafe.Pointer, w *Writer) {
@@ -443,16 +432,10 @@ func (*stringCodec) Encode(ptr unsafe.Pointer, w *Writer) {
 
 type bytesCodec struct {
 	sliceType *reflect2.UnsafeSliceType
-	convert   func(*Reader) []byte
 }
 
 func (c *bytesCodec) Decode(ptr unsafe.Pointer, r *Reader) {
-	var b []byte
-	if c.convert != nil {
-		b = c.convert(r)
-	} else {
-		b = r.ReadBytes()
-	}
+	b := r.ReadBytes()
 	c.sliceType.UnsafeSet(ptr, reflect2.PtrOf(b))
 }
 
@@ -582,18 +565,12 @@ func (c *timeMicrosCodec) Encode(ptr unsafe.Pointer, w *Writer) {
 var one = big.NewInt(1)
 
 type bytesDecimalCodec struct {
-	prec    int
-	scale   int
-	convert func(*Reader) []byte
+	prec  int
+	scale int
 }
 
 func (c *bytesDecimalCodec) Decode(ptr unsafe.Pointer, r *Reader) {
-	var b []byte
-	if c.convert != nil {
-		b = c.convert(r)
-	} else {
-		b = r.ReadBytes()
-	}
+	b := r.ReadBytes()
 	if i := (&big.Int{}).SetBytes(b); len(b) > 0 && b[0]&0x80 > 0 {
 		i.Sub(i, new(big.Int).Lsh(one, uint(len(b))*8))
 	}
@@ -634,19 +611,12 @@ func (c *bytesDecimalCodec) Encode(ptr unsafe.Pointer, w *Writer) {
 }
 
 type bytesDecimalPtrCodec struct {
-	prec    int
-	scale   int
-	convert func(*Reader) []byte
+	prec  int
+	scale int
 }
 
 func (c *bytesDecimalPtrCodec) Decode(ptr unsafe.Pointer, r *Reader) {
-	var b []byte
-	if c.convert != nil {
-		b = c.convert(r)
-	} else {
-		b = r.ReadBytes()
-	}
-
+	b := r.ReadBytes()
 	if i := (&big.Int{}).SetBytes(b); len(b) > 0 && b[0]&0x80 > 0 {
 		i.Sub(i, new(big.Int).Lsh(one, uint(len(b))*8))
 	}